models/impossible-travel-detector/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.0

# Geographic calculations
geopy>=2.4.0


jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
